ADD .../xa/arc/zzz/xaxzexport* (2)
2 files:
+ urt/src/phpc/xa/arc/zzz/xaxzexport.php

This is on the phpc (CLI) side, refered to by previous commit.
This function is an intermediate to the next commit:
urt/src/php_/xa/arc/zzz/_xaxzexport.php

+ urt/txt/fb/xa/arc/zzz/xaxzexport-en.jsun
The feedback table used by the above, for help and error messages.

NOTE:
Classes and functions inside urt/src/* are in the scope of the
multilingual textual feedback system making it easy to provide
feedback to the user.
This function generates feedback and guidance on the command line,
which is why this code is not inside urt/bin/*